+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_localizations/flutter_localizations.dart';
 import 'pages/heatmap_calendar_example.dart';
 import 'pages/heatmap_example.dart';
 
@@ -14,6 +15,12 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Heatmap Demo',
+      localizationsDelegates: GlobalMaterialLocalizations.delegates,
+      locale: const Locale('es', ''),
+      supportedLocales: const [
+        Locale('es', ''),
+        Locale('en', ''),
+      ],
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),