+++ lib/src/data/constants.dart
@@ -2,3 +2,6 @@ const double kDefaultBlockSizeCalendar = 42.0;
 const double kDefaultBlockSizeMap = 20.0;
 const double kDefaultSpacingBlock = 6.0;
 const double kDefaultSpacingTip = 6.0;
+const String kDefaultDatePatternCalendar = 'MMMM y';
+const int kDefaultStartDayOfWeek = 7;
+const int kDefaultDaysOfWeek = 7;