+++ lib/src/util/datasets_util.dart
@@ -4,15 +4,16 @@ import './date_util.dart';
 class DatasetsUtil {
   /// Filtering [datasets] where the key is on the same month of [referenceDate].
   static Map<DateTime, int> filterMonth(
-      Map<DateTime, int>? datasets, DateTime referenceDate) {
-    return Map.from(datasets ?? {})
-      ..removeWhere(
-        (date, value) =>
-            !(date.isAfter(DateUtil.startDayOfMonth(referenceDate)) &&
-                    date.isBefore(DateUtil.endDayOfMonth(referenceDate)) ||
-                date == DateUtil.endDayOfMonth(referenceDate) ||
-                date == DateUtil.startDayOfMonth(referenceDate)),
-      );
+    Map<DateTime, int>? datasets,
+    DateTime referenceDate,
+  ) {
+    return Map.from(datasets ?? {})..removeWhere(
+      (date, value) =>
+          !(date.isAfter(DateUtil.startDayOfMonth(referenceDate)) &&
+                  date.isBefore(DateUtil.endDayOfMonth(referenceDate)) ||
+              date == DateUtil.endDayOfMonth(referenceDate) ||
+              date == DateUtil.startDayOfMonth(referenceDate)),
+    );
   }
 
   /// Get maximum value of [datasets].