+++ lib/src/widget/heatmap_week_text.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import '../util/date_util.dart';
 
 class HeatMapWeekText extends StatelessWidget {
   /// The margin value for correctly space between labels.
@@ -11,14 +10,23 @@ class HeatMapWeekText extends StatelessWidget {
   /// The TextStyle of week labels.
   final TextStyle? textStyle;
 
-  const HeatMapWeekText({super.key, this.margin, this.textStyle, this.size});
+  /// The list of labels for week days.
+  final List<String> weekDayLabels;
+
+  const HeatMapWeekText({
+    super.key,
+    this.margin,
+    this.textStyle,
+    this.size,
+    required this.weekDayLabels,
+  });
 
   @override
   Widget build(BuildContext context) {
     return Column(
       mainAxisSize: MainAxisSize.min,
       children: <Widget>[
-        for (String label in DateUtil.WEEK_LABEL)
+        for (String label in weekDayLabels)
           Container(
             height: size ?? 20,
             margin: margin ?? const EdgeInsets.all(2.0),